+++ globals_for_gh.nim
@@ -1,0 +1,8 @@
+## Shared constants for gh_nimrod_doc_pages.
+##
+## For project information see https://github.com/gradha/gh_nimrod_doc_pages.
+
+const
+  md_extensions* = [".md", ".markdown"]
+  rst_extensions* = [".rst"]
+  html_extensions* = [".html", ".htm"]